+++ src/Model/vpg_generation_option.cpp
@@ -166,7 +166,8 @@ std::shared_ptr<Json> VPGGenerationOption::ToJson() const
         json->AddString(L"ManagerTypeDirectory", _ManagerTypeDirectory);
         json->AddString(L"ObjectTypeDirectory", _ObjectTypeDirectory);
 
-        json->AddString(L"ObjectDirectory", _ObjectDirectory);
+        json->AddString(L"ObjectDirectoryHpp", _ObjectDirectoryHpp);
+        json->AddString(L"ObjectDirectoryCpp", _ObjectDirectoryCpp);
         json->AddString(L"PropertyAccessorDirectoryHpp", _PropertyAccessorDirectoryHpp);
         json->AddString(L"PropertyAccessorDirectoryCpp", _PropertyAccessorDirectoryCpp);
 
@@ -237,8 +238,10 @@ void VPGGenerationOption::DeserializeJson(std::shared_ptr<IDocument> document) c
         if (json->IsContainKey(L"ObjectTypeDirectory"))
             this->SetObjectTypeDirectory(json->GetString(L"ObjectTypeDirectory"));
 
-        if (json->IsContainKey(L"ObjectDirectory"))
-            this->SetObjectDirectory(json->GetString(L"ObjectDirectory"));
+        if (json->IsContainKey(L"ObjectDirectoryHpp"))
+            this->SetObjectDirectoryHpp(json->GetString(L"ObjectDirectoryHpp"));
+        if (json->IsContainKey(L"ObjectDirectoryCpp"))
+            this->SetObjectDirectoryCpp(json->GetString(L"ObjectDirectoryCpp"));
         if (json->IsContainKey(L"PropertyAccessorDirectoryHpp"))
             this->SetPropertyAccessorDirectoryHpp(json->GetString(L"PropertyAccessorDirectoryHpp"));
         if (json->IsContainKey(L"PropertyAccessorDirectoryCpp"))